backend/pyproject.toml
[project]
name = "atss-backend"
version = "0.1.0"
description = "ATSS Matarat Backend"
authors = [
    {name = "Antigravity",email = "antigravity@google.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-multipart (>=0.0.21,<0.0.22)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
